requirements.txt
Django
djangorestframework
drf-yasg
djangorestframework-simplejwt
django-filter